requirements.txt
setuptools
ZODB3
lxml==3.4.0
zope.interface
python-dateutil
Chameleon
pyduktape
